+++ tests/test_correct_bound.py
@@ -14,14 +14,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-"""Test correctness of the Cython implementation of PVAL-BOUND."""
+"""Tests for Cython implementation of the XL-mHG bounds."""
 
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
-from builtins import *
-
-import sys
-import itertools as it
+from builtins import str as text
 
 import numpy as np
 from scipy.stats import hypergeom