+++ xlmhg/__init__.py
@@ -2,6 +2,7 @@ from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
 from builtins import *
 
-from xlmhg.test import xlmhg_test
+from .result import mHGResult
+from .test import xlmhg_test, get_xlmhg_test_result
 
-__all__ = ['xlmhg_test']
+__all__ = ['mHGResult', 'xlmhg_test', 'get_xlmhg_test_result']